ADN-cursos-back/java-arquitectura-hexagonal/microservicio/infraestructura/src/test/java/com/ceiba/alumno/controlador/ComandoControladorAlumnoTest.java
package com.ceiba.alumno.controlador;

import com.ceiba.ApplicationMock;
import com.ceiba.alumno.comando.ComandoAlumno;
import com.ceiba.alumno.servicio.testdatabuilder.ComandoAlumnoTestDataBuilder;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.annotation.DirtiesContext;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.Matchers.hasSize;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@ExtendWith(SpringExtension.class)
@WebMvcTest(ComandoControladorAlumno.class)
@ContextConfiguration(classes= ApplicationMock.class)
@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
public class ComandoControladorAlumnoTest {

    @Autowired
    private ObjectMapper objectMapper;

    @Autowired
    private MockMvc mocMvc;

    @Test
    @DisplayName("Debería crear un alumno")
    void deberiaCrearUnAlumno() throws Exception{
        // arrange
        ComandoAlumno alumno = new ComandoAlumnoTestDataBuilder().build();
        // act - assert
        mocMvc.perform(post("/alumnos")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(objectMapper.writeValueAsString(alumno)))
                .andExpect(status().isOk())
                .andExpect(content().json("{'valor': 2}"));
    }

    @Test
    @DisplayName("Debería actualizar un alumno")
    void deberiaActualizarUnAlumno() throws Exception{
        // arrange
        Long id = 1L;
        ComandoAlumno alumno = new ComandoAlumnoTestDataBuilder().build();
        // act - assert
        mocMvc.perform(put("/alumnos/{id}",id)
                        .contentType(MediaType.APPLICATION_JSON)
                        .content(objectMapper.writeValueAsString(alumno)))
                .andExpect(status().isOk());
    }

    @Test
    @DisplayName("Debería eliminar un alumno")
    void deberiaEliminarUnAlumno() throws Exception {
        // arrange
        Long id = 1L;
        // act - assert
        mocMvc.perform(delete("/alumnos/{id}",id)
                        .contentType(MediaType.APPLICATION_JSON)
                        .accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk());

        mocMvc.perform(get("/alumnos")
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", hasSize(0)));
    }
}
